modify: rename `genshin_app` -> `hoyolab`
