+++ apps/genshin/__init__.py
@@ -2,7 +2,7 @@ from .browser import *
 from .checks import *
 from .enka import *
 from .find_codes import *
-from .genshin_app import *
+from .hoyolab import *
 from .leaderboard import *
 from .tasks import *
 from .utility import *
