+++ apps/genshin/hoyolab.py
